refactor: add index.js to component folders
